fix: return to original db style
- modularizing sucks. too many errors
